cambio carpetas

+++ Libft/sources/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: agomez-m <agomez-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/12 13:34:15 by agomez-m          #+#    #+#             */
-/*   Updated: 2023/09/12 13:34:56 by agomez-m         ###   ########.fr       */
+/*   Updated: 2023/09/12 19:09:21 by agomez-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int ft_strlen(const char *s)
+size_t ft_strlen(const char *s)
 {
-    int i;
+    size_t i;
 
     i = 0;
     while (s[i])
